+++ client/index.js
@@ -2,7 +2,7 @@ var viewport = require('./viewport');
 var school = require('./school');
 var sub = require('./sub');
 var messages = require('./messages');
-var schoolSync = require('./school-sync');
+var tankSync = require('./tank-sync');
 var animate = require('./animate');
 var interaction = require('./interaction');
 //var PIXI = require('pixi.js');
@@ -19,7 +19,7 @@ var HEIGHT = boundingRect.height;
 school.init();
 sub.init();
 messages.init();
-schoolSync.init();
+tankSync.init();
 animate.load(function() {
   interaction.init();
 });
