+++ client/tank-sync.js
